small error in photo url corrected
git-svn-id: svn+ssh://foss.ulster.ac.uk/svn/opus/trunk@642 54dc3e3f-df17-4cd7-99f1-b1610cd3c088

+++ html/controllers/supervisor/index_home.php
@@ -142,7 +142,7 @@
 
   function display_photo(&$waf, &$user)
   {
-    $username = (int) WA::request("username");
+    $username = WA::request("username");
     $fullsize = WA::request("fullsize");
     require_once("model/Photo.class.php");
